backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Supabase & Auth
supabase==2.3.0
PyJWT[crypto]==2.8.0
python-jose[cryptography]==3.3.0

# File handling
aiofiles==23.2.1
zipfile36==0.1.3

# PCB parsing - KiCad
# Note: pcbnew comes with KiCad installation, can't pip install
# For development, we'll use file parsing approach
pyparsing==3.1.1

# Gerber parsing
pygerber==2.2.0

# PDF generation
reportlab==4.0.7
Pillow==10.1.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# HTTP client
httpx==0.24.1
requests==2.31.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/OpenAI
openai>=1.0.0

# KiCad parsing (S-expression parser)
sexpdata>=0.0.3
